+++ Strategies/Exploration.cpp
@@ -13,7 +13,7 @@ Exploration::Exploration(GameManager& gm, string nom)
     // La distance du npc à la tuile
     // La distance moyenne de cette tuile aux autres tuiles qui seront visités !
     // Le degré d'intêret de la tuile. 
-void Exploration::saveScore(MapTile tile, float cout, Npc& npc, vector<int> tilesAVisiter) noexcept {
+void Exploration::saveScore(MapTile tile, int cout, Npc& npc, vector<int> tilesAVisiter) noexcept {
     float score = 0;
 
     // Si on a déjà visité cette case, son score est nul
